refactor-ir-cake: IRContext moved to sigma.compiler
